Moved unit test files to differentiate with upcoming tests
